[IMP] ..company_aware..r: test server functions
