+++ mail_company_aware_sender/tests/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import test_company_aware_sender
+from . import test_mail_server
